KAS-2622 rename 2 files + finalize part 3
