+++ cypress/selectors/newsletter.selectors.js
