+++ cypress/support/commands/agenda-commands.js
@@ -8,7 +8,7 @@ import agenda from '../../selectors/agenda.selectors';
 import form from '../../selectors/form.selectors';
 import modal from '../../selectors/modal.selectors';
 import utils from '../../selectors/utils.selectors';
-import auComponents from '../../selectors/au-component-selectors';
+import auComponents from '../../selectors/au-component.selectors';
 import dependency from '../../selectors/dependency.selectors';
 import route from '../../selectors/route.selectors';
